Highlight overdue days in not drawn loan alerts
There is a 10 day grace period where lenders can record the initial draw
on a loan after the 6 month deadline has passed. This wasn't apparent
on the existing loan alert chart. This adds a fourth group to that
chart, highlighting the loans that are overdue.

This includes a large refactoring of loan alerts so they are composed of
any number of groups. Each group defines its priority and date range.
This allows for greater flexibility in grouping and rendering alerts,
rather than being fixed to high/medium/low priority over a set date
range.