Allow CfeUser's to perform a loan status amendment
A number of loans have been identified as ineligible for EFG. These have
been drawn and in some cases repaid or a claim has been made.

Claims made will be refunded via a recovery but depending on the reason
for the ineligibility, the claim limit should not be reduced. This is to
prevent the lender freeing up further claim limit when they are not
entitled to it.

This is accomplished through a status amendment on a loan, performed by
a CfeUser. A status amendment can be made on any loan that has been
guaranteed.

See examples at https://git.io/vVbWa