Update zeroed out created_at and updated_at values
Database records in various tables that existed prior to migrating to
this app had their `created_at` and `updated_at` set to 0000-00-00
00:00:00.

When the extractor runs it converts these datetime to the current
datetime (i.e. the time the extractor runs). This causes problems when
working with the data locally, like when querying data by creation
date.

This adds a data migration which updates all those zeroed out values to
the midnight of the morning when this app was launched.