Update premium schedule for fixed repayment amount
Outstanding balance for each quarter is the number of months repaid
so far multiplied by the fixed repayment amount.

E.g. £1,000 per month x 15 months

Premium is not charged in the final quarter if the repayment duration is
not evenly divisible by 3. E.g. 31 month loan. This is because the
premium is only charged on the outstanding balance at the beginning of
the each quarter. For a 31 month loan, the balance at month 33 is £0.