Only allow lender login when not on new legal agreement
Only users of lenders who have not yet signed the legal agreement can
use the Legacy App. This adds a new boolean on lenders which is managed
by CfeAdmins. When this boolean is enabled lender admins and lender
users for this lender are redirected to a page telling them to use the
new Master App.

When the new legal agreement is not signed, they can use the app as
normal.