Remove default values for secrets
We very nearly had a security incident with EFG because of the way we add
initializers during deployment. We ended up almost using the public version
of the `secret_key` in production.

To stop this from happening, the app must read it's config from the
environment. This may mean that you have to adjust your development
environments to provide these variables.