Premium schedule capital repayment holiday fix
Fixed repayment amount premium schedules were not properly catering for
the capital repayment holiday. The loan term should be extended by
the number of repayment holiday months. Once the holiday is over,
then capital begins to be repaid at the fixed amount.

So a loan with a value of £120,000, a fixed monthly repayment of
£10,000 and a capital repayment holiday of 6 months, will have a
term of 18 months ((£120,000 / £10,000) + 6).