Fix change password log out issues
Fixes an intermittent issue on production. When changing their password,
a user is sometimes logged out after submitted the form. The new 
password is not saved and they have to log back in with their old 
password.

Moving the `sign_in` to after the user record is saved seems to fix it
but I cannot reproduce this failure in development or in a test.